Set LimitMaxbytes type to uint64.
The memcached protocol defines the `limit_maxbytes` as '32u'. But in
real test it may exceed 4294967295 and cause error. So it's better set
LimitMaxbytes to uint64 for better compatibility.

+++ memcache/memcache.go
@@ -206,7 +206,7 @@ type GeneralStats struct {
     Reclaimed uint64
     BytesRead uint64
     BytesWritten uint64
-    LimitMaxbytes uint32
+    LimitMaxbytes uint64
     Threads uint32
     ConnYields uint64
     HashPowerLevel uint32
